sum-storage works; tests separated
