+++ pallets/sum-storage/Cargo.toml
@@ -6,15 +6,15 @@ edition = "2018"
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
-sp-std = { rev = '3e651110aa06aa835790df63410a29676243fc54', git = 'https://github.com/paritytech/substrate.git', default_features = false}
-sp-runtime = { rev = '3e651110aa06aa835790df63410a29676243fc54', git = 'https://github.com/paritytech/substrate.git', default_features = false}
-frame-support = { rev = '3e651110aa06aa835790df63410a29676243fc54', git = 'https://github.com/paritytech/substrate.git', default_features = false}
-frame-system = { rev = '3e651110aa06aa835790df63410a29676243fc54', git = 'https://github.com/paritytech/substrate.git', default_features = false}
+sp-std = { tag = 'v2.0.0-alpha.1', git = 'https://github.com/paritytech/substrate.git', default_features = false}
+sp-runtime = { tag = 'v2.0.0-alpha.1', git = 'https://github.com/paritytech/substrate.git', default_features = false}
+frame-support = { tag = 'v2.0.0-alpha.1', git = 'https://github.com/paritytech/substrate.git', default_features = false}
+frame-system = { tag = 'v2.0.0-alpha.1', git = 'https://github.com/paritytech/substrate.git', default_features = false}
 
 
 [dev-dependencies]
-sp-io = { rev = '3e651110aa06aa835790df63410a29676243fc54', git = 'https://github.com/paritytech/substrate.git', default_features = false}
-sp-core = { rev = '3e651110aa06aa835790df63410a29676243fc54', git = 'https://github.com/paritytech/substrate.git', default_features = false}
+sp-io = { tag = 'v2.0.0-alpha.1', git = 'https://github.com/paritytech/substrate.git', default_features = false}
+sp-core = { tag = 'v2.0.0-alpha.1', git = 'https://github.com/paritytech/substrate.git', default_features = false}
 
 [features]
 default = ["std"]
