+++ pallets/sum-storage/rpc/runtime-api/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Joshy Orndorff"]
 edition = "2018"
 
 [dependencies]
-sp-api = { rev = '3e651110aa06aa835790df63410a29676243fc54', git = 'https://github.com/paritytech/substrate.git', default_features = false}
+sp-api = { tag = 'v2.0.0-alpha.1', git = 'https://github.com/paritytech/substrate.git', default_features = false}
 
 [dev-dependencies]
 serde_json = "1.0.41"
